requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1

# Async and networking
aiohttp==3.9.1
websockets==12.0
httpx==0.25.2

# Exchange API
python-binance==1.0.19
ccxt==4.2.25

# Data processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
numba==0.58.1

# Technical analysis
# ta-lib==0.4.28  # Закомментировано - сложная установка на Windows
pandas-ta==0.3.14b0

# Machine Learning
xgboost==2.0.3
lightgbm==4.2.0
scikit-learn==1.3.2
optuna==3.5.0  # For hyperparameter optimization

# Database
redis==5.0.1
psycopg2-binary==2.9.9  # PostgreSQL/TimescaleDB
# questdb==1.0.1  # Закомментировано - опциональная зависимость
sqlalchemy==2.0.25
alembic==1.13.1  # Database migrations

# Monitoring and logging
prometheus-client==0.19.0
structlog==24.1.0
colorama==0.4.6
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0

# Utils
python-dateutil==2.8.2
pytz==2023.3
orjson==3.9.10  # Fast JSON
marshmallow==3.20.1
dacite==1.8.1  # Dataclass utils

# System monitoring
psutil==5.9.6

# Deployment and Performance
# uvloop==0.19.0  # Fast event loop - только для Unix систем
gunicorn==21.2.0
python-daemon==3.0.1

# Job scheduling
apscheduler==3.10.4